refactor(drivers): extract request_pairing_token to shared v2 module
Move the bridge pairing helper from opencode.rs into v2/mod.rs as a pub
function so Claude, Kimi, and Codex drivers can reuse it in Phase 2.
Update opencode.rs call site to super::request_pairing_token and move
the three HTTP integration tests to mod.rs where they test the shared impl.
