+++ src/agent/drivers/v2/opencode.rs
@@ -3,8 +3,6 @@
 use std::io::Write;
 use std::process::{Command, Stdio};
 use std::sync::{Arc, Mutex};
-use std::time::Duration;
-
 use anyhow::{bail, Context};
 use async_trait::async_trait;
 use tokio::io::{AsyncBufReadExt, BufReader};
@@ -54,40 +52,6 @@ fn build_mcp_chat_config(
             ],
         }),
     }
-}
-
-/// Request a one-time pairing token from the shared bridge's admin endpoint.
-///
-/// The bridge mints the token, binds it to `agent_key`, and we embed it in
-/// the MCP URL written to `opencode.json`. On any failure (network, non-2xx,
-/// malformed response) this bubbles up — the driver surfaces the error to
-/// the caller rather than silently falling back to the stdio path.
-async fn request_pairing_token(bridge_endpoint: &str, agent_key: &str) -> anyhow::Result<String> {
-    let base = bridge_endpoint.trim_end_matches('/');
-    let url = format!("{base}/admin/pair");
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(10))
-        .build()
-        .context("failed to build HTTP client for bridge pairing")?;
-    let res = client
-        .post(&url)
-        .json(&serde_json::json!({ "agent_key": agent_key }))
-        .send()
-        .await
-        .with_context(|| format!("bridge unreachable at {url}"))?;
-    if !res.status().is_success() {
-        let status = res.status();
-        let body = res.text().await.unwrap_or_default();
-        anyhow::bail!("bridge-pair failed: {status} {body}");
-    }
-    let json: serde_json::Value = res
-        .json()
-        .await
-        .context("invalid bridge pairing response (not JSON)")?;
-    json["token"]
-        .as_str()
-        .map(|s| s.to_string())
-        .ok_or_else(|| anyhow::anyhow!("missing 'token' in bridge response: {json}"))
 }
 
 // ---------------------------------------------------------------------------
@@ -289,7 +253,7 @@ impl AgentHandle for OpencodeHandle {
         // fails we surface the error — no silent fallback to stdio, so
         // misconfiguration is loud.
         let pairing_token = if let Some(endpoint) = &self.spec.bridge_endpoint {
-            Some(request_pairing_token(endpoint, &self.key).await.with_context(
+            Some(super::request_pairing_token(endpoint, &self.key).await.with_context(
                 || format!("failed to pair with bridge at {endpoint} for agent {}", self.key),
             )?)
         } else {
@@ -898,88 +862,4 @@ mod tests {
         assert_eq!(config["type"], "local");
     }
 
-    // ---- request_pairing_token HTTP integration ----
-
-    async fn spawn_mock_bridge(
-        response: axum::response::Response,
-    ) -> (String, tokio::task::JoinHandle<()>) {
-        use axum::routing::post;
-        use axum::Router;
-        use std::sync::Mutex as StdMutex;
-
-        let shared = Arc::new(StdMutex::new(Some(response)));
-        let app = Router::new().route(
-            "/admin/pair",
-            post(move || {
-                let shared = shared.clone();
-                async move {
-                    shared
-                        .lock()
-                        .unwrap()
-                        .take()
-                        .expect("mock bridge only answers once")
-                }
-            }),
-        );
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let port = listener.local_addr().unwrap().port();
-        let url = format!("http://127.0.0.1:{port}");
-        let handle =
-            tokio::spawn(async move { axum::serve(listener, app).await.unwrap() });
-        // Give the listener a tick to be ready.
-        tokio::time::sleep(std::time::Duration::from_millis(25)).await;
-        (url, handle)
-    }
-
-    #[tokio::test]
-    async fn request_pairing_token_parses_success_response() {
-        let response = axum::response::Response::builder()
-            .status(200)
-            .header("content-type", "application/json")
-            .body(axum::body::Body::from(
-                serde_json::json!({"token": "tok-123"}).to_string(),
-            ))
-            .unwrap();
-        let (url, handle) = spawn_mock_bridge(response).await;
-
-        let token = request_pairing_token(&url, "agent-abc").await.unwrap();
-        assert_eq!(token, "tok-123");
-        handle.abort();
-    }
-
-    #[tokio::test]
-    async fn request_pairing_token_surfaces_non_2xx() {
-        let response = axum::response::Response::builder()
-            .status(500)
-            .body(axum::body::Body::from("boom"))
-            .unwrap();
-        let (url, handle) = spawn_mock_bridge(response).await;
-
-        let err = request_pairing_token(&url, "agent-abc")
-            .await
-            .expect_err("non-2xx should bubble up");
-        let msg = format!("{err:#}");
-        assert!(msg.contains("bridge-pair failed"), "got: {msg}");
-        handle.abort();
-    }
-
-    #[tokio::test]
-    async fn request_pairing_token_errors_on_missing_field() {
-        let response = axum::response::Response::builder()
-            .status(200)
-            .header("content-type", "application/json")
-            .body(axum::body::Body::from(
-                serde_json::json!({"wrong_field": "x"}).to_string(),
-            ))
-            .unwrap();
-        let (url, handle) = spawn_mock_bridge(response).await;
-
-        let err = request_pairing_token(&url, "agent-abc")
-            .await
-            .expect_err("missing token field should fail");
-        let msg = format!("{err:#}");
-        assert!(msg.contains("missing 'token'"), "got: {msg}");
-        handle.abort();
-    }
 }
